requirements.txt
aiogram==3.13.1
aiofiles==24.1.0
aiosqlite==0.20.0
cryptography==43.0.1
openai==1.51.2
pandas==2.2.3
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.35
httpx==0.27.0
ddgs==9.11.1
tavily-python==0.7.22
tqdm==4.66.5
